requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.3.0=py37haa95532_0
aiohttp=3.8.1=py37h2bbff1b_1
aiosignal=1.2.0=pyhd3eb1b0_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py37haa95532_0
asynctest=0.13.0=py_0
attrs=21.4.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py37haa95532_0
bottleneck=1.3.5=py37h080aedc_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py37h2bbff1b_1003
ca-certificates=2022.4.26=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
cairo=1.16.0=h63a05c6_1001
certifi=2022.6.15=py37haa95532_0
cffi=1.15.1=py37h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py37haa95532_0
colorama=0.4.5=py37haa95532_0
cryptography=3.4.8=py37h71e12ea_0
cycler=0.11.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
docopt=0.6.2=py37haa95532_1
et_xmlfile=1.1.0=py37haa95532_0
expat=2.4.9=h6c2663c_0
fftw=3.3.9=h2bbff1b_1
flatbuffers=22.10.26=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
fribidi=1.0.10=h8d14728_0
frozenlist=1.2.0=py37h2bbff1b_0
gast=0.3.3=py_0
getopt-win32=0.1=h8ffe710_0
glib=2.69.1=h5dc1a3c_1
google-auth=1.28.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
graphite2=1.3.14=hd77b12b_1
graphviz=2.50.0=hdb8b0d4_0
grpcio=1.42.0=py37hc60d5dd_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
gts=0.7.6=h69fa6b1_0
h5py=2.10.0=py37h5e291fa_0
harfbuzz=4.3.0=hda2c7e1_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2022.1.0=h6049295_2
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py37haa95532_0
intel-openmp=2021.4.0=haa95532_3556
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_0
keras-applications=1.0.8=py_1
keras-nightly=2.12.0.dev2022103007=pypi_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.2=py37hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=pypi_0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_4
libgd=2.3.3=ha43c60c_1
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h23ce68f_0
libtiff=4.4.0=h8a3f274_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markdown=3.3.4=py37haa95532_0
matplotlib=3.5.1=py37haa95532_1
matplotlib-base=3.5.1=py37hd77b12b_1
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py37h2bbff1b_0
mkl_fft=1.3.1=py37h277e83a_0
mkl_random=1.2.2=py37hf11a4ad_0
multidict=6.0.2=py37h2bbff1b_0
munkres=1.1.4=py_0
numexpr=2.8.3=py37hb80d3ca_0
numpy=1.21.5=py37h7a0a035_3
numpy-base=1.21.5=py37hca35cd5_3
oauthlib=3.1.1=pyhd3eb1b0_0
openpyxl=3.0.9=pyhd3eb1b0_0
openssl=1.1.1q=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.3.5=py37h6214cd6_0
pango=1.50.7=h78c2152_0
pcre=8.45=hd77b12b_0
pillow=9.2.0=py37hdc2b20a_1
pip=22.2.2=py37haa95532_0
pixman=0.38.0=hfa6e2cd_1003
ply=3.11=py37_0
protobuf=3.19.6=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pydot=1.4.2=py37h03978a9_2
pyjwt=2.4.0=py37haa95532_0
pyopenssl=19.1.0=py37_0
pyparsing=3.0.9=py37haa95532_0
pyqt=5.15.7=py37hd77b12b_0
pyqt5-sip=12.11.0=py37hd77b12b_0
pyreadline=2.1=py37_1
pysocks=1.7.1=py37_1
python=3.7.13=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytz=2021.3=pyhd3eb1b0_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h3ad3cdb_4
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.0.2=py37hf11a4ad_1
scipy=1.7.3=py37h7a0a035_2
seaborn=0.11.2=pyhd3eb1b0_0
setuptools=65.4.0=py37haa95532_0
sip=6.6.2=py37hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.3=h2bbff1b_0
tb-nightly=2.11.0a20221030=pypi_0
tensorboard=2.8.0=py37haa95532_0
tensorboard-data-server=0.6.0=py37haa95532_0
tensorboard-plugin-wit=1.8.1=py37haa95532_0
tensorflow-base=2.3.0=eigen_py37h17acbac_0
tensorflow-estimator=2.3.0=pyheb71bc4_0
tensorflow-io-gcs-filesystem=0.27.0=pypi_0
termcolor=1.1.0=py37haa95532_1
tf-estimator-nightly=2.12.0.dev2022103008=pypi_0
tf-nightly=2.12.0.dev20221030=pypi_0
tf-nightly-intel=2.12.0.dev20221030=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.2=py37h2bbff1b_0
typing-extensions=4.3.0=py37haa95532_0
typing_extensions=4.3.0=py37haa95532_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.0.2=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py37haa95532_0
wincertstore=0.2=py37haa95532_2
wrapt=1.13.3=py37h2bbff1b_2
xz=5.2.6=h8cc25b3_0
yarl=1.8.1=py37h2bbff1b_0
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
